File structure refactor for Wayat
