+++ .gitignore
@@ -1,3 +1,7 @@
+WAYAT/windows/
+WAYAT/linux/
+WAYAT/macos/
+
 # Miscellaneous
 *.class
 *.lock
@@ -153,3 +157,4 @@ SPIKES/Routing/Navigation/windows/flutter/generated_plugins.cmake
 SPIKES/State Management/MobX&GetIt/ios/Runner.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
 SPIKES/State Management/MobX&GetIt/ios/Runner.xcodeproj/project.xcworkspace/xcshareddata/WorkspaceSettings.xcsettings
 WAYAT/.env
+*.cc
