+++ WAYAT/lib/features/create_event/page/create_event_page.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:wayat/features/create_event/controller/create_event_controller.dart';
+
+class CreateEventPage extends StatefulWidget {
+  const CreateEventPage({Key? key}) : super(key: key);
+
+  @override
+  State<CreateEventPage> createState() => _CreateEventPageState();
+}
+
+class _CreateEventPageState extends State<CreateEventPage> {
+  final CreateEventController controller = CreateEventController();
+
+  @override
+  void dispose() {
+    controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      child: Form(
+        child: Column(
+          children: const [],
+        ),
+      ),
+    );
+  }
+}
