+++ WAYAT/lib/features/event_detailed_view/page/event_details_page.dart
