+++ WAYAT/lib/features/home/pages/home_provisional.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class HomeProvPage extends StatelessWidget {
+  const HomeProvPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('HOME'),
+      ),
+    );
+  }
+}
