--- WAYAT/lib/pages/home_provisional.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-
-class HomeProvPage extends StatelessWidget {
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Text('HOME'),
-     ),
-   );
-  }
-}
