+++ WAYAT/lib/services/contact/contact_service.dart
@@ -1,4 +1,4 @@
-import 'package:wayat/domain/contact.dart';
+import 'package:wayat/domain/contact/contact.dart';
 import 'package:wayat/services/service.dart';
 
 abstract class ContactService extends Service {
